requirements.txt
telepot==12.7
python-dotenv==1.0.0
fyers-apiv2==2.0.5 
pandas==1.3.5